FRO-04: refactoring in progress
